cleanup network service
